ホットキー関連の命令をplugin_browserに移動
